Fix deprecation warnings in tests, reorganize test utilities a little and add some comments
